+++ src/Model/Table/VideosCategoriesTable.php
@@ -108,6 +108,7 @@ class VideosCategoriesTable extends AppTable {
             'foreignKey' => 'category_id'
         ]);
 		
+        $this->addBehavior('Timestamp');
         $this->addBehavior('MeCms.Tree');
     }