+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # 2.x branch
+## 2.3 branch
+### 2.3.0
+* now the sitemap uses the cache and handles `lastmod` and `priority` values;
+* now videos categories have "created" and "modified" fields;
+* rewrote the code to generate the backend menus.
+
 ## 2.2 branch
 ### 2.2.1
 * added functions to generate the site sitemap.